WIP: trying to make extended draw tools work
